+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 0.3.0
+
+- New option `--delete-project` to delete al project versions from a docat server
+- New option `--system-certs` to use the system SSL certificate bundle instead of [certifi](https://github.com/certifi/python-certifi).
+
 ## Version 0.2.0
 
 - New option `--verbose` for some debug information